toollama/WORKING/requirements.txt
annotated-types==0.7.0
anthropic==0.46.0
anyio==4.8.0
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
filelock==3.17.0
Flask==3.1.0
Flask-Cors==5.0.0
fsspec==2025.2.0
google==3.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
mlx==0.21.1
mlx-lm==0.20.6
numpy==2.2.3
openai==1.64.0
packaging==24.2
pillow==11.1.0
protobuf==5.29.3
pydantic==2.10.6
pydantic_core==2.27.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
sentencepiece==0.2.0
sniffio==1.3.1
soupsieve==2.6
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
urllib3==2.3.0
waybackpy==3.0.6
Werkzeug==3.1.3
